paddleocr_service/requirements.txt
fastapi
uvicorn
paddleocr[all]
numpy
Pillow
opencv-python-headless
httpx